requirements.txt
Flask==2.1.0
google-cloud-storage==1.31.0
google-auth==1.21.0
six==1.15.0
DateTime
gunicorn==20.1.0